backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
mysql-connector-python==8.1.0
python-dotenv==1.0.1
jinja2==3.1.2
python-multipart==0.0.6
pandas==2.2.2
openpyxl==3.1.2

